WL#3072 Maria Recovery * added replaying of REDO_REPAIR_TABLE, but disabled it as mysterious linker errors appear. * after replaying RENAME/REPAIR, we must bump create_rename_lsn for idempotency of maria_read_log.
sql/mysqld.cc:
  typo
storage/maria/ma_checkpoint.c:
  silence compiler warning
storage/maria/ma_recovery.c:
  * added replaying of REDO_REPAIR_TABLE, but disabled it as
  mysterious linker errors appear.
  * after replaying RENAME/REPAIR, we must bump create_rename_lsn
  for idempotency of maria_read_log.

+++ storage/maria/ma_checkpoint.c
@@ -48,7 +48,7 @@
   Background page flushing is not used.
   So, needed pagecache functions for doing this flushing are not yet pushed.
 */
-#define flush_pagecache_blocks_with_filter(A,B,C,D,E) (((int)D) * 0)
+#define flush_pagecache_blocks_with_filter(A,B,C,D,E) (int)(((ulong)D) * 0)
 /**
    filter has to return 0, 1 or 2: 0 means "don't flush this page", 1 means
    "flush it", 2 means "don't flush this page and following pages".
